refactor #167 : 히스토리 조회 모달 content 분리 및 Suspense, Error 적용
